requirements.txt
black==22.1.0
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.2
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
python-dotenv==0.19.2
requests==2.27.1
tomli==2.0.1
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
